requirements.txt
yfinance==0.2.54
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
hmmlearn==0.3.2
scikit-learn==1.5.0
gymnasium==0.29.1
stable-baselines3==2.3.2
tensorboard==2.16.2
torch==2.3.0
anthropic==0.86.0
matplotlib==3.9.0
plotly==5.22.0
jupyter==1.0.0
pytest==8.2.0
pyarrow==16.1.0
streamlit==1.38.0
python-dotenv==1.0.1
cython==3.0.10
